Finishing up my 1.0 I think
Added the ability to change the top view controller.
Updated the code to be device non-specific, it now works in all orientations and all devices
Added a new variable to set how open the slide view should be
All animations now use a setState method to set the state and animate

Need to add velocity towards the speed of the animation but overall I'm pretty happy with the way it animates now.  All animations are based on the size of the view and new variable and are not hard coded.